backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.3
bs4==0.0.1
beautifulsoup4==4.11.1
requests==2.28.1
psycopg2-binary
celery==5.2.7
redis==4.3.4
django-celery-results==2.4.0
celery-progress==0.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0


